Create variableDef.iss
I have completed most of the variable declarations for later development. Includes the registry and environment variables. The registry and environment variables return the parameter values through functions. If it does not conform to the later requirements, it can be revised further.

However, for WiseScript 185 to 206 lines of content has not yet completed the scripting language conversion, it can be supplemented according to the later development process.