Update variableDef.iss
Instead of just using a string '%MAINDIR%, I have modify the variables MAINDIR and BACKUP. I am not sure it is right, we only can find out it is right or wrong when we use these variables in the future. If there is someting wrong, let me know as soon as posisbly.